add definition for DynamicQuantizeLSTM
