+++ onnxruntime/core/graph/contrib_ops/contrib_defs.h
@@ -28,6 +28,7 @@ namespace contrib {
 void RegisterContribSchemas();
 void RegisterNchwcSchemas();
 void RegisterNhwcSchemas();
+void RegisterQuantizationSchemas();
 
 constexpr const float kDefaultSkipLayerNormEpsilon = 1e-12f;
 constexpr const float kDefaultEmbedLayerNormEpsilon = 1e-12f;
